requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
deepdiff>=6.0
python-slugify>=8.0.0
pyyaml>=6.0
